Add files via upload
This program is an object-oriented implementation of Abstract Base Class graph where I have showed two different types of graph implementation (adjacency matrix and adjacency list) and how much they create difference on sparse and dense graph. This program also covers the issue of deciding which presentation should be used for any graph depending on the internal representation of a graph.